+++ log.log
@@ -140,3 +140,5 @@ The last packet sent successfully to the server was 0 milliseconds ago. The driv
 04/07/2018 15:08:00	|	null		|	Cancelar Aula
 04/07/2018 15:08:51	|	null		|	Cadastrar Aula
 04/07/2018 15:09:00	|	null		|	Cancelar Aula
+04/07/2018 15:35:03	|	null		|	Cadastrar Aula
+04/07/2018 15:35:05	|	null		|	Cancelar Aula
