+++ src/com/topicos/aluno/AlunoDAO.java
@@ -1,0 +1,4 @@
+package com.topicos.aluno;
+
+public class AlunoDAO {
+}
