+++ src/com/topicos/professor/ProfessorDAO.java
@@ -1,0 +1,4 @@
+package com.topicos.professor;
+
+public class ProfessorDAO {
+}
